Added nearest_neighbor(), changed some function names, changed insert()
Renamed the search_with_vectors -> search() to just mirror FAISS index.search()

Created nearest_neighbor() which will fetch records from the database and use
those vectors to pass to search() to find the nearest records in the database
to those of interest.  This does not return yourself in the neighbors list.

insert() now adds vectors to FAISS index if the index exist and then writes back
out to disk to change it.  Probably want to handle versioning at some point, but
that seems complicated right now.