Added sync_index_to_database() & train_index()
Added these to make things easier if someone wants to train on a GPU.
This way they can do just the training on a GPU they can save it afterwards
and then copy the file to a different machine to add the vectors.

For now, I'm assuming that people that want to use the GPU can figure it
out for themselves how to transfer the needed parts to the GPU resourse.