More GPU support & updated to v0.2.2
Added the attach() to make it easy to attach a FAISS index to the
VekterDB. Also now check if the index is a GpuIndex before saving
to disk. If it is, then call faiss.index_gpu_to_cpu(self.index)
to move it to the CPU before saving. Otherwise FAISS throws an error